[dtMediaWiki] move mediawikiapi to lib folder
