+++ contrib/dtMediaWiki/dtMediaWiki.lua
@@ -8,8 +8,8 @@ Dependencies:
 * lua-multipart-post: HTTP Multipart Post helper
 ]]
 local dt = require "darktable"
-local MediaWikiApi = require "contrib/dtMediaWiki/mediawikiapi"
-local version = 46
+local MediaWikiApi = require "contrib/dtMediaWiki/lib/mediawikiapi"
+local version = 47
 
 --[[The version number is generated by .git/hooks/pre-commit (+x)
     with the following content:
