+++ contrib/dtMediaWiki/lib/mediawikiapi.lua
